Reorganise AppInfo
