+++ BannerText/AppInfo.h
@@ -5,4 +5,4 @@
 #define VER_PRODUCTNAME_STR         "BannerText"
 #define VER_FILE_DESCRIPTION_STR    "Generate and display a Text Banner in the console"
 
-#include "..\Common\AppInfo.h"
+#include "../Common/AppInfo.h"
