+++ Common/AppInfo.h
@@ -1,10 +1,16 @@
 #pragma once
 
-// ReSharper disable IdentifierTypo
-
 #include "BuildInfo.h"
 #include "Version.h"
 #include "Macros.h"
+
+#include "../DNX.App/DNXAppDetails.h"
+
+using namespace DNX::App;
+
+// ReSharper disable IdentifierTypo
+// ReSharper disable CppInconsistentNaming
+// ReSharper disable CppClangTidyModernizeUseEqualsDefault
 
 #define VER_COMPANY_NAME            "DNX Solutions Ltd"
 #define VER_COPYRIGHT_STR           "Copyright © 2018-" TO_STR(COPYRIGHT_YEAR) " " VER_COMPANY_NAME
@@ -33,3 +39,18 @@
 #define VER_FILEOS                  VOS_NT_WINDOWS32
 #define VER_FILEFLAGS               VER_VER_DEBUG
 #define VER_FILETYPE                VFT_APP
+
+
+//------------------------------------------------------------------------------
+// AppInfo
+class AppInfo final : public AppDetails
+{
+public:
+    AppInfo()
+    {
+        Name        = VER_PRODUCTNAME_STR;
+        Description = VER_FILE_DESCRIPTION_STR;
+        Version     = VER_FILE_VERSION_STR;
+        Copyright   = VER_COPYRIGHT_STR;
+    }
+};
